fix: Centralize package versions, fix build errors from v0.3.0 merge
Resolve NU1605 package downgrade (System.Threading.Channels 9.x vs 10.0.1
required by DotCompute 0.5.3), centralize 22 hardcoded package versions
across 12 project files to use Directory.Build.props MSBuild variables,
fix Interlocked round-robin off-by-one in AdaptiveLoadBalancer, fix
long-to-int cast in DotComputeMemoryAllocator, restore CS8669 suppression
for DotCompute generator output, and align HLC tests to millisecond
granularity of packed timestamp format.

+++ src/Orleans.GpuBridge.Backends.DotCompute/Memory/DotComputeMemoryAllocator.cs
@@ -230,7 +230,7 @@ internal sealed class DotComputeMemoryAllocator : IMemoryAllocator
             TotalBytesAllocated: totalAllocated,
             TotalBytesInUse: totalInUse,
             TotalBytesFree: totalAllocated - totalInUse,
-            AllocationCount: allocationCount,
+            AllocationCount: (int)allocationCount,
             FreeBlockCount: _memoryPools.Count,
             LargestFreeBlock: largestFreeBlock,
             FragmentationPercent: fragmentationPercent,
